Use a webworker to run the AI
Run teh AI in a webworker to remove the UI lag when searching for
a best move.

Due to the object structure the grid needs to be syncronised between
the worker and the main thread, this is a small overhead.

On my laptop this removes completly the flicker when doing moves.